// src/estadisticas/dto/juez-completo.dto.ts
export class JuezCompletoDto {
  n_id_juez: number;
  l_activo: string;
  usuario_id: number;
  n_id_juez_tipo_id: number;
  x_juez_tipo_descripcion: string;
  x_nombres: string;
  x_app_paterno: string;
  x_app_materno: string;
  x_dni: string;
  x_telefono: string;
  email: string;
  username: string;
  profile_image: string;
  l_mensaje: boolean;
  n_id_sexo_id: number;
  x_sexo_descripcion: string;
  n_instancia_id: number;
  x_nom_instancia: string;
}